Implemented XML Layouts and Majority of the functionality
Implemented the XML layouts of the 2 pages needed. Along with the XML for the row to show the data on the first page.

Implemented class to hold the data of Todo Items.

Implemented the Adapter to show the data in the list view.

Implemented a Merge Sort class with the sorting and merging functions using an array of TodoItem class.

Implemented the DBHelper class to use the SQLite DB, this is where further testing needs to be done as to begin Inserting to the DB and Table worked but now when gathering the data it does not get multiple amounts and only returns one.