Commented all Java files, All Requirements are working
Commented all the Java classes and files.

Ability to create a new todo item - done.

List existing todo items in a list, ordered by completion date and displaying each of the todo item’s properties (the description may be ellipsized) - done

Able to mark a todo item as completed, with a mark on the todo item in the list to signify completion - done

Delete a todo item - done

Tap a todo item to open the details page - done

Display all the todo item’s properties - done

Edit the properties such as title, description, change date - done

Mark as completed and return to the list - done

Delete and return to the list - done

SL