Began Fixing Issues
Fixed the issue with crashing when deleting a fresh new entry as none will be able to delete causing a crash in Details Activity. Fix was removing the Delete Button view when a new entry is being made.

Began to change to an Array List and will be altering the merge sort class to work with an Array List instead of standard TodoItem array.

Also looked into doing a comparator instead of using the merge sort to make it simpler but this makes for certain API's to be used whereas I might as well do it myself.

Removed Images folder and the images within as the vector library was enough for the basic functionality needed.

SL