Added Time Picker
Added Time Picker to the details page XML, along with linking it within the java class of Details Activity.

Created a new class for the merge sorting algorithm using Array List, also need to create the adapter to use the array list instead of the array and test the merge sort of the array list.

SL