Working on Refreshing ListView
Working on the ListView within the Details Activity to have it refresh whenever a user adds or removes from the database and the data list along with updating correctly when a user select is complete switch on both pages.

Also altered code in Main Activity to use the merge sort correctly with the array list instead of using the old merge sort with standard arrays and altered the on click listener and the adapter to use the array list as well.

SL